requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
PyMySQL==1.1.1
python-dotenv==1.0.1
Authlib==1.5.2
cryptography==42.0.7
Flask-Migrate==4.1.0